refactor: updateMutation 로직 분리
